Compute relative humidity indicators from grouped data

The function computes relative humidity indicators from grouped data. Expects relative humidity in percentage.

Usage

summarise_rel_humidity(.x, value_var, normals_df)

Arguments

.x

grouped data, created with dplyr::group_by()

value_var

name of the variable with temperature values.

normals_df

normals data, created with summarise_normal()

Value

A tibble.

Details

The dry and wet spells indicators are computed based on climatological normals, created with the summarise_normal() function and passed with the normals_df argument. Keys to join the normals data must be present (like id, year, and month) and use the same names.

The following indicators are computed for each group.

count Count of data points

normal Climatological normal, from normals_df argument

mean Average

median Median

sd Standard deviation

se Standard error

max Maximum value

min Minimum value

p10 10th percentile

p25 25th percentile

p75 75th percentile

p90 90th percentile

dry_spells_3d Count of dry spells occurences, with 3 or more consecutive days with relative humidity bellow the climatological normal value minus 10 percent

dry_spells_5d Count of dry spells occurences, with 5 or more consecutive days with relative humidity bellow the climatological normal value minus 10 percent

wet_spells_3d Count of wet spells occurences, with 3 or more consecutive days with relative humidity above the climatological normal value plus 10 percent

wet_spells_5d Count of wet spells occurences, with 5 or more consecutive days with relative humidity above the climatological normal value plus 10 percent

dry_days Count of dry days, when the relative humidity is bellow the 10th percentile

wet_days Count of wet days, when the relative humidity is above the 90th percentile

h_21_30 Count of days with relative humidity between 21% and 30%. Attention level

h_12_20 Count of days with relative humidity between 12% and 20%. Alert level

h_11 Count of days with relative humidity bellow 12%. Emergence level

Examples

# Compute monthly normals
normals <- rel_humidity_data |>
  # Identify year
  dplyr::mutate(year = lubridate::year(date)) |>
  # Identify month
  dplyr::mutate(month = lubridate::month(date)) |>
  # Group by id variable, year and month
  dplyr::group_by(code_muni, year, month) |>
  summarise_normal(date_var = date, value_var = value, year_start = 1961, year_end = 1990) |>
  dplyr::ungroup()

# Compute indicators
rel_humidity_data |>
 # Identify year
 dplyr::mutate(year = lubridate::year(date)) |>
 # Identify month
 dplyr::mutate(month = lubridate::month(date)) |>
 # Group by id variable, year and month
 dplyr::group_by(code_muni, year, month) |>
 # Compute relative humidity indicators
 summarise_rel_humidity(value_var = value, normals_df = normals) |>
 # Ungroup
 dplyr::ungroup()
#> # A tibble: 1,440 × 24
#>    code_muni  year month count normal  mean median    sd    se   max   min   p10
#>        <int> <dbl> <dbl> <int>  <dbl> <dbl>  <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1   3106200  1961     1    31   88.0  88.0   88.8  5.84 1.05   96.0  72.9  80.9
#>  2   3106200  1961     2    28   80.6  80.6   80.1  6.73 1.27   93.7  68.7  73.3
#>  3   3106200  1961     3    31   78.9  78.9   79.0  5.80 1.04   89.8  64.6  71.5
#>  4   3106200  1961     4    30   76.1  76.1   74.9  5.92 1.08   87.4  67.3  69.6
#>  5   3106200  1961     5    31   76.8  76.8   76.8  6.29 1.13   91.7  63.6  70.1
#>  6   3106200  1961     6    30   75.6  75.6   76.0  3.71 0.678  83.7  68.4  71.5
#>  7   3106200  1961     7    31   69.7  69.7   69.9  4.90 0.880  83.0  59.5  62.7
#>  8   3106200  1961     8    31   61.0  61.0   60.5  4.30 0.773  70.8  51.7  56.8
#>  9   3106200  1961     9    30   57.6  57.6   56.8  5.21 0.952  72.5  48.5  52.3
#> 10   3106200  1961    10    31   65.3  65.3   63.3  9.49 1.71   89.6  52.3  56.8
#> # ℹ 1,430 more rows
#> # ℹ 12 more variables: p25 <dbl>, p75 <dbl>, p90 <dbl>, dry_spells_3d <int>,
#> #   dry_spells_5d <int>, wet_spells_3d <int>, wet_spells_5d <int>,
#> #   dry_days <int>, wet_days <int>, h_21_30 <int>, h_12_20 <int>, h_11 <int>