Compute windspeed indicators from grouped data

The function computes windspeed (u2) indicators from grouped data. Expects windspeed in meters per second (m/s).

Usage

summarise_windspeed(.x, value_var, normals_df)

Arguments

.x

grouped data, created with dplyr::group_by()

value_var

name of the variable with temperature values.

normals_df

normals data, created with summarise_normal()

Value

A tibble.

Details

The high and low u2 indicators are computed based on climatological normals, created with the summarise_normal() function and passed with the normals_df argument. Keys to join the normals data must be present (like id, year, and month) and use the same names.

The following indicators are computed for each group.

count Count of data points

normal Climatological normal, from normals_df argument

mean Average

median Median

sd Standard deviation

se Standard error

max Maximum value

min Minimum value

p10 10th percentile

p25 25th percentile

p75 75th percentile

p90 90th percentile

l_eto_3 Count of sequences of 3 days or more with evapotranspirations bellow the climatological normal

l_eto_5 Count of sequences of 5 days or more with evapotranspirations bellow the climatological normal

h_eto_3 Count of sequences of 3 days or more with evapotranspirations above the climatological normal

h_eto_5 Count of sequences of 5 days or more with evapotranspirations above the climatological normal

Examples

# Compute monthly normals
normals <- windspeed_data |>
  # Identify year
  dplyr::mutate(year = lubridate::year(date)) |>
  # Identify month
  dplyr::mutate(month = lubridate::month(date)) |>
  # Group by id variable, year and month
  dplyr::group_by(code_muni, year, month) |>
  summarise_normal(date_var = date, value_var = value, year_start = 1961, year_end = 1990) |>
  dplyr::ungroup()

# Compute indicators
windspeed_data |>
 # Identify year
 dplyr::mutate(year = lubridate::year(date)) |>
 # Identify month
 dplyr::mutate(month = lubridate::month(date)) |>
 # Group by id variable, year and month
 dplyr::group_by(code_muni, year, month) |>
 # Compute windspeed indicators
 summarise_windspeed(value_var = value, normals_df = normals) |>
 # Ungroup
 dplyr::ungroup()
#> # A tibble: 1,440 × 19
#>    code_muni  year month count normal  mean median    sd     se   max   min
#>        <int> <dbl> <dbl> <int>  <dbl> <dbl>  <dbl> <dbl>  <dbl> <dbl> <dbl>
#>  1   3106200  1961     1    31  1.14  1.14   1.15  0.378 0.0679  1.77 0.324
#>  2   3106200  1961     2    28  0.934 0.934  0.920 0.303 0.0572  1.42 0.435
#>  3   3106200  1961     3    31  1.33  1.33   1.21  0.635 0.114   3.25 0.384
#>  4   3106200  1961     4    30  1.17  1.17   1.14  0.340 0.0621  2.04 0.489
#>  5   3106200  1961     5    31  1.01  1.01   1.00  0.352 0.0632  1.71 0.255
#>  6   3106200  1961     6    30  1.37  1.37   1.43  0.569 0.104   2.53 0.524
#>  7   3106200  1961     7    31  1.27  1.27   1.19  0.365 0.0655  2.12 0.762
#>  8   3106200  1961     8    31  1.41  1.41   1.45  0.392 0.0705  2.17 0.674
#>  9   3106200  1961     9    30  1.35  1.35   1.21  0.607 0.111   2.94 0.499
#> 10   3106200  1961    10    31  1.98  1.98   2.06  0.547 0.0982  2.86 0.935
#> # ℹ 1,430 more rows
#> # ℹ 8 more variables: p10 <dbl>, p25 <dbl>, p75 <dbl>, p90 <dbl>, l_u2_3 <int>,
#> #   l_u2_5 <int>, h_u2_3 <int>, h_u2_5 <int>